requirements.txt
asn1crypto==0.24.0
azure-common==1.1.23
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
boto3==1.9.253
botocore==1.12.253
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
croniter==0.3.30
cryptography==2.8
dask==2.6.0
distributed==2.6.0
docker==4.1.0
docutils==0.15.2
fsspec==0.5.2
future==0.18.1
HeapDict==1.0.1
idna==2.8
ijson==2.5.1
jmespath==0.9.4
locket==0.2.0
marshmallow==3.2.1
marshmallow-oneofschema==2.0.1
msgpack==0.6.2
mypy-extensions==0.4.3
numpy==1.17.3
oauthlib==3.1.0
pandas==0.25.2
partd==1.0.0
pendulum==2.0.5
prefect==0.7.3
psutil==5.6.3
pycparser==2.19
pycryptodomex==3.9.0
PyJWT==1.7.1
pyOpenSSL==19.0.0
python-box==3.4.5
python-dateutil==2.8.0
python-slugify==4.0.0
pytz==2019.3
pytzdata==2019.3
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
six==1.12.0
snowflake-connector-python==1.9.1
snowflake-sqlalchemy==1.1.15
sortedcontainers==2.1.0
SQLAlchemy==1.3.10
tabulate==0.8.5
tblib==1.5.0
text-unidecode==1.3
toml==0.10.0
toolz==0.10.0
tornado==6.0.3
typing==3.7.4.1
typing-extensions==3.7.4.1
urllib3==1.25.6
websocket-client==0.56.0
zict==1.0.0
